refactor: create folders for context in use cases
