+++ backend/src/main/java/br/ufrn/imd/pastora/controllers/MonitorController.java
@@ -4,10 +4,10 @@ import br.ufrn.imd.pastora.controllers.dto.CreateHttpMonitorDto;
 import br.ufrn.imd.pastora.persistence.MonitorModel;
 import br.ufrn.imd.pastora.persistence.repository.MonitorRepository;
 import br.ufrn.imd.pastora.persistence.repository.ServiceRepository;
-import br.ufrn.imd.pastora.usecases.CreateMonitorUseCase;
-import br.ufrn.imd.pastora.usecases.GetMonitorUseCase;
-import br.ufrn.imd.pastora.usecases.GetMonitorsByServiceUseCase;
-import br.ufrn.imd.pastora.usecases.GetMonitorsUseCase;
+import br.ufrn.imd.pastora.usecases.monitor.CreateMonitorUseCase;
+import br.ufrn.imd.pastora.usecases.monitor.GetMonitorUseCase;
+import br.ufrn.imd.pastora.usecases.monitor.GetMonitorsByServiceUseCase;
+import br.ufrn.imd.pastora.usecases.monitor.GetMonitorsUseCase;
 import br.ufrn.imd.pastora.utils.AuthenticatedUserUtils;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
