+++ backend/src/main/java/br/ufrn/imd/pastora/controllers/ServiceController.java
@@ -5,13 +5,13 @@ import br.ufrn.imd.pastora.domain.Service;
 import br.ufrn.imd.pastora.exceptions.BusinessException;
 import br.ufrn.imd.pastora.persistence.ServiceModel;
 import br.ufrn.imd.pastora.persistence.repository.ServiceRepository;
-import br.ufrn.imd.pastora.usecases.CreateServiceUseCase;
-import br.ufrn.imd.pastora.usecases.DeleteServiceUseCase;
-import br.ufrn.imd.pastora.usecases.GetServiceByNameText;
-import br.ufrn.imd.pastora.usecases.GetServiceIconUseCase;
-import br.ufrn.imd.pastora.usecases.GetServiceUseCase;
-import br.ufrn.imd.pastora.usecases.GetServicesUseCase;
-import br.ufrn.imd.pastora.usecases.UpdateServiceUseCase;
+import br.ufrn.imd.pastora.usecases.service.CreateServiceUseCase;
+import br.ufrn.imd.pastora.usecases.service.DeleteServiceUseCase;
+import br.ufrn.imd.pastora.usecases.service.GetServiceByNameText;
+import br.ufrn.imd.pastora.usecases.service.GetServiceIconUseCase;
+import br.ufrn.imd.pastora.usecases.service.GetServiceUseCase;
+import br.ufrn.imd.pastora.usecases.service.GetServicesUseCase;
+import br.ufrn.imd.pastora.usecases.service.UpdateServiceUseCase;
 import br.ufrn.imd.pastora.utils.AuthenticatedUserUtils;
 import lombok.AllArgsConstructor;
 import lombok.SneakyThrows;
