+++ backend/src/main/java/br/ufrn/imd/pastora/domain/ExecutionData.java
@@ -25,7 +25,6 @@ public class ExecutionData {
     }
 
     public enum Status {
-        IDLE,
         RUNNING,
         FINISHED,
     }
