+++ backend/src/main/java/br/ufrn/imd/pastora/domain/MonitorHttpDefinition.java
@@ -17,8 +17,6 @@ public class MonitorHttpDefinition  {
     @NotNull
     String url;
 
-    String type = "HTTP";
-
     @NotNull
     HttpMethod method;
