+++ backend/src/main/java/br/ufrn/imd/pastora/usecases/monitor/GetMonitorUseCase.java
@@ -1,4 +1,4 @@
-package br.ufrn.imd.pastora.usecases;
+package br.ufrn.imd.pastora.usecases.monitor;
 
 import java.util.Optional;
