+++ backend/src/main/java/br/ufrn/imd/pastora/usecases/service/GetServiceUseCase.java
@@ -1,4 +1,4 @@
-package br.ufrn.imd.pastora.usecases;
+package br.ufrn.imd.pastora.usecases.service;
 
 import java.util.Optional;
